requirements.txt
asgiref==3.7.2
asttokens==2.4.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==5.0.3
django-redis
executing==2.0.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
ipython==8.23.0
isodate==0.6.1
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
pytest-mock==3.12.0
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
stripe
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
forex-python
celery
django-celery-beat
redis
drf_yasg
djangorestframework
django-cors-headers
djangorestframework-simplejwt
